requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
openai>=1.0.0
chromadb>=0.4.0
firecrawl-py>=0.0.16
unstructured>=0.10.0
python-dotenv>=1.0.0 